+++ src/app/api/auth/[...nextauth]/route.ts
@@ -1,85 +1,6 @@
-import { NextAuthOptions } from "next-auth";
-import CredentialsProvider from "next-auth/providers/credentials";
-import bcrypt from "bcryptjs";
-import dbConnenct from "@/lib/dbConnect";
-import UserModel from "@/model/User";
+import NextAuth from "next-auth";
+import { nextAuthOptions } from "./options";
 
-export const nextAuthOptions: NextAuthOptions = {
-  providers: [
-    CredentialsProvider({
-      id: "credentials", // not present in docs but used in various examples
-      name: "Credentials",
-      credentials: {
-        // input fields for the credentials provider
-        email: { label: "Email", type: "text" },
-        password: { label: "Password", type: "password" },
-      },
+const handler = NextAuth(nextAuthOptions);
 
-      async authorize(credentials: any): Promise<any> {
-        await dbConnenct();
-
-        const { email, username } = credentials.identifier;
-
-        try {
-          const user = await UserModel.findOne({
-            // used or here for future functionality of username login
-            $or: [{ email }, { username }],
-          });
-
-          if (!user) {
-            throw new Error("No user found with this email");
-          }
-
-          if (!user.isVerified) {
-            throw new Error("Please verify your email to login");
-          }
-
-          const isPasswordValid = await bcrypt.compare(
-            // compare the password with the password in the database
-            credentials.password,
-            user.password
-          );
-
-          if (isPasswordValid) {
-            return user;
-          } else {
-            throw new Error("Incorrect password");
-          }
-        } catch (err: any) {
-          throw new Error(err);
-        }
-      },
-    }),
-  ],
-
-  callbacks: {
-    async jwt({ token, user }) {
-      if (user) {
-        token._id = user._id.toString(); // Modified token to store additional user data
-        token.isVerified = user.isVerified;
-        token.isAcceptingMessages = user.isAcceptingMessages;
-        token.username = user.username;
-      }
-      return token;
-    },
-    async session({ session, token }) {
-      if (token) {
-        session.user._id = token._id; // Modified token to store additional user data
-        session.user.isVerified = token.isVerified;
-        session.user.isAcceptingMessages = token.isAcceptingMessages;
-        session.user.username = token.username;
-      }
-      return session;
-    },
-  },
-
-  pages: {
-    signIn: "/sign-in",
-  },
-
-  session: {
-    strategy: "jwt",
-  },
-
-  secret: process.env.NEXTAUTH_SECRET,
-};
+export { handler as GET, handler as POST };
